Add source branch capability to GitHub example
Each new branch in a GitHub repo is created from a source branch, which
defaults to `main`. Previously the GitHub Terraform example did not have
any way to change the source branch. This commit will enable source
branches to be specified with "from" in the branch name, like
`"new-branch-name from source-branch-name"`, in the list of branches in
`var.repos`.